requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
influxdb-client==1.36.1
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
requests==2.28.2
schedule==1.2.0
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
